backend_service/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
Flask-Bcrypt
Flask-Migrate
alembic
python-dotenv
PyJWT
pandas
requests
gunicorn
psycopg2-binary
